Move why3 specific code to backend::*
